Add @RegisterBank macros (#2)
Adds a pair of macros used to define a bank of registers directly in
Swift code called `@RegisterBank` and `@RegisterBank(offset:)`. Referred
to as bank and offset macro respectively.

The bank macro converts a struct defining the layout of nested registers
and nested register banks into a reference type containing the base
address of the structure. The offset macro converts stored properties
into computed properties offset from the base address introduced by the
bank macro. The bank and offset macros emit a number of diagnostics and
fix-its to aid programmers.

Adds a variety of test cases for each macro individually and both used
in conjunction.