Fix macro attribute signature (#3)
Fixes a bug where the `@RegisterBank(offset:)` macro was not usable
because the argument labels of the macro declaration did not match the
argument labels expected by the macro definition, leading to an error
diagnostic at time of use.

Fixes a bug where `RegisterBankOffsetMacro` was not exposed by the MMIO
macro compiler plugin.