Add initial support for discontiguous bit fields
Updates the bit field macros 'bits' argument to accept a list of bit
ranges. The register macro uses this information to generate a bit field
descriptor type conforming to either ContiguousBitField or
DiscontiguousBitField. The public api for users remains unchanged.

This commit doesn't introduce any diagnostics for overlapping bit fields
within a discontiguous bit fields. A future commit will include those
diagnostics as well as additional diagnostics for overlapping bit fields
across fields in a register.

In the future the implementation of bit field descriptors could be
improved. At a high level we would introduce a BitField protocol with
two requirements: least significant bit and most significant bit. We
would then extend FixedWidthInteger with a subscript taking a variadic
list of these fields. This change would allow us to remove the
distinction between contiguous and discontiguous bit fields. In order to
make this change we need variadic pack iteration to land and we need to
be able to use variadic packs without using runtime metadata.