Build placeholder with Syntax tree instead of String (#8)
Refactor generation of ParsableMacro AttributeSyntax with argument
placeholders. This code previously built a String then asked
swift-syntax to parse it into a tree, this led to an internal error due
to the use of editor placeholders in the attribute. We can avoid this
error by building the attribute using swift syntax types directly and
get a small performance win.