Add projected bit fields (#27)
Removes raw access to bit fields from generated Register Read and Write
views. Users will now need to use the `raw` computed property of the
Read and Write views to access bit fields as the underlying register
storage integer type.

For example, if a programmer previously had a register with the
following field and mutation

```swift
@Register(bitWidth: 32)
struct R32 {
  @ReadWrite(bits: 0..<1)
  var en: EN
}

r32.modify { $0.en = 1 }
```

The modify closure must be updated to: `$0.raw.en = 1`

---

Introduces the `BitFieldProjectable` protocol which describes how types
can be encoded to or decoded from raw bit fields. Bit field macros now
have an optional `as:` argument which take a type conforming to the
`BitFieldProjectable` protocol and expose the field as that type.
Extending the above example with this new functionality we get the
following snippet:

```swift
@Register(bitWidth: 32)
struct R32 {
  @ReadWrite(bits: 0..<1, as: Bool.self)
  var en: EN
}

r32.modify { $0.en = true }
```

Field and projected type bit widths cannot be statically validated to
match due to limitations with the Swift language. However MMIO
guarantees that such uses compile to trap if the projection is used.