RawRepresentable BitFieldProjectable conformance
Adds conditional extension `RawRepresentable` types whose `RawValue` is
a `FixedWidthInteger` to make it easier to adopt `BitFieldProjectable`.
These types will now get default implementations for the
`init(storage:)` and `storage(_:)` requirements and only need to provide
an implementation for the `bitWidth: Int { get }` requirement.

Conforming a `RawRepresentable` type to `BitFieldProjectable` is only
valid if all bit representations of the raw value within the declared
bit width are valid values.

Fixes #32