Improve diagnostics for macro annotated members (#64)
Updates the emitted error for members missing a macro annotation to
include fix its. If a member can be annotated with more than one macro,
a fix it to insert each macro will be emitted.

Removes dead code to emit an unused old diagnostic.

Fixes a bug where `Register.unsafeAddress` was declared as a `let`
instead of a `var`.