Add simple in process FileCheck alternative (#78)
This commit adds a minimal implementation of LLVM's FileCheck which runs
in process. This allows us to run the swift-mmio FileCheck tests in CI
and makes it easier for contributors to run a full test suite without
needing an LLVM toolchain build. MMIOFileCheckTests will still prefer to
use LLVM FileCheck if it can be found on the host machine. Alternatively
the 'SWIFT_MMIO_USE_SIMPLE_FILECHECK' environment variable can be set to
force the use of the in process implementation.