Skip locking package build directory
SwiftPM 6 introduces a lock in the package build directory to avoid
concurrent builds/tests. Unfortuntely this behavior is needed by the
MMIOFileCheckTests currently. This commit adds a flag to avoid locking
the package build directory when building with SwiftPM 6.

Additionally, I have an inflight PR to change `--ignore-lock true` to
`--ignore-lock` so this commit will require a follow-up fix, after that
change in SwiftPM is merged (apple/swift-package-manager#7384).