Remove IR used in SVD2Swift (#107)
SVD2Swift previously leveraged a simplified IR to avoid directly
operating on SVD types. This allowed it to perform mutations to the
underlying data structure when converting from SVD types to IR types.

The new version of SVD2Swift directly operates on SVD types so that a
common core set of API can be used by SVD2LLDB and SVDExplorer instead
of needing an IR.

This commit does the following things to achieve the goal of removing
the IR. 

- SVD2Swift export functions are moved from IR types to SVD types.
- An inflate function is added to handle derived-from relationships.
- Common Swift extensions in SVD2Swift are moved to MMIOUtilities. 

The end result is a simpler implementation of SVD2Swift and fixes for
SVD2LLDB to walk through derived-from relationships.

A future improvement should be made so that inflation is lazy and
request-based so the various tools do not need to inflate the entire
tree if the user only interacts with a subset of the tree.