Add Projection associated type to BitField (#130)
Updates the `BitField` associated type to require a `Projection`
associated type which conforms to `BitFieldProjectable`. Renames
existing `insert` and `extract` methods to `insertBits` and
`extractBits` respectively. Adds new typed variants of `insert` and
`extract` which operate on `Projection`.

The result of this change is that `preconditionMatchingBitWidth` can be
hidden from the MMIO module's public API and the macro expansion code is
simpler.