Rewrite parsing code to avoid Sendable issues (#156)
The previous implementation of the parser used Sendable closures and
passed values and types into the closures. This has issues with the
updated isolated conformance models in Swift 6.2. The parser has been
rewritten with concrete types to avoid the Sendable issues.

Ideally all of these code could be deleted in favor of a parser library
in swiftlang.