Always expose interposer API
The interposer API is fairly undiscoverable and the docc catalog does
not include the symbols because they are hidden behind a `#if`. This has
the knock on affect of not being able to link to the symbols too.

This commit changes MMIO to always expose interposing API. The APIs are
marked with a deprecation warning when not compiled with
`FEATURE_INTERPOSABLE` and are a no-op. As a result the docc catalog now
includes documentation for these symbols and can reference them
throughout the catalog.