Enable iOS builds (#165)
The SVD target has been preventing iOS and derivative OS builds because
XMLNode and XMLElement are not present in the relevant SDKs. This commit
rewrites the XML layer of SVD to use NSXMLParser instead of XMLNode and
XMLElement to build up an intermediate tree and then creates the SVD
types from that tree. In the future it may be possible to construct the
SVD types directly without the intermediate tree.